feat: Add Fabric 1.21 support with new mixins, render helper, and build configuration updates.

+++ fabric-1.21.9/src/main/java/com/criticalrange/mixins/extra/MixinWindow.java
@@ -21,7 +21,7 @@ public class MixinWindow {
     private int height;
 
     @Inject(method = "setScaleFactor", at = @At("HEAD"))
-    private void vulkanmodExtra$optimizeMacResolution(double d, CallbackInfo ci) {
+    private void vulkanmodExtra$optimizeMacResolution(int d, CallbackInfo ci) {
         // Check if macOS resolution reduction is enabled
         if (VulkanModExtra.CONFIG.extraSettings.reduceResolutionOnMac) {
             // Check if we're running on macOS
